fix: move wdio to TS
