+++ apps/desktop/e2e/tests/index.spec.js
@@ -1,6 +1,6 @@
 describe('GitButler', () => {
 	it('should have the root element', async () => {
 		const element = await $('body.text-base');
-		expect(element).toExist();
+		await expect(element).toExist();
 	});
 });
