+++ apps/desktop/wdio.conf.ts
@@ -9,13 +9,13 @@ let tauriDriver: ChildProcess;
 export const config: Options.WebdriverIO = {
 	hostname: '127.0.0.1',
 	port: 4444,
-	specs: ['./e2e/tests/**/*.spec.js'],
+	specs: ['./e2e/tests/**/*.spec.js', './e2e/tests/**/*.spec.ts'],
 	maxInstances: 1,
 	capabilities: [
 		{
 			// @ts-expect-error custom tauri capabilities
+			maxInstances: 1,
 			'tauri:options': {
-				maxInstances: 1,
 				application: '../../target/release/git-butler-dev'
 			}
 		}
@@ -45,6 +45,7 @@ export const config: Options.WebdriverIO = {
 
 	waitforTimeout: 10000,
 	connectionRetryTimeout: 120000,
+	connectionRetryCount: 0,
 
 	// ensure we are running `tauri-driver` before the session starts so that we can proxy the webdriver requests
 	beforeSession: () =>
